admin/requirements/local.txt
-r base.txt

faker==17.6.0
flake8==4.0
django-debug-toolbar==3.4.0
pytest==7.2.2
watchgod==0.8.2
